refactor(bin): extract allocator and scheduler into separate classes
This came with a minor performance loss (<5%) which I could recover my comparing perf traces to the previous commit.
