+++ bin/simexp/event_queue.cpp
@@ -1,142 +1,6 @@
 #include "event_queue.hpp"
 #include "fmt/base.h"
 
-inline u64 index_to_bitmask(u8 index) {
-  if (index >= 64) [[unlikely]]
-    throw std::out_of_range("Index must be less than 64");
-  return 1ULL << index;
-}
-
-void EventLoop::Allocator::free_event(Event *ev) {
-  // Ensure that the event actually belongs to us.
-  if (const void *address = ev; address == nullptr) [[unlikely]]
-    return;
-  size_t slot_index = (reinterpret_cast<std::byte *>(ev) - _slots[0].data) / sizeof(EventSlot);
-  free_event(slot_index);
-}
-
-void EventLoop::Allocator::free_event(u8 idx) {
-  if (idx >= MAX_EVENTS) [[unlikely]]
-    throw std::runtime_error("Index too high");
-  else if (!_slots_used.test(idx)) [[unlikely]] throw std::runtime_error("Event slot is not currently allocated");
-  std::destroy_at((Event *)_slots[idx].data);
-
-  _counters.freed++;
-  // TODO: how to communicate to scheduler that event is freed??
-  //_scheduled.clear_bit(idx);
-  //_event_dependencies[idx].clear(), _event_dependents[idx].clear();
-  _slots_used.clear_bit(idx);
-}
-
-Event *EventLoop::Allocator::at(u8 idx) {
-  if (idx >= MAX_EVENTS) [[unlikely]]
-    throw std::out_of_range("Index must be less than MAX_EVENTS");
-  else if (!_slots_used.test(idx)) [[unlikely]] throw std::runtime_error("Event slot is not currently allocated");
-  return reinterpret_cast<Event *>(_slots[idx].data);
-}
-
-const Event *EventLoop::Allocator::at(u8 idx) const {
-  if (idx >= MAX_EVENTS) [[unlikely]]
-    throw std::out_of_range("Index must be less than MAX_EVENTS");
-  else if (!_slots_used.test(idx)) [[unlikely]] throw std::runtime_error("Event slot is not currently allocated");
-  return reinterpret_cast<const Event *>(_slots[idx].data);
-}
-
-u8 EventLoop::Scheduler::count() const { return _queue_size; }
-
-u64 EventLoop::Scheduler::current_tick() const noexcept { return _current_tick; }
-
-bool EventLoop::Scheduler::skip(u64 ticks) {
-  if (count() == 0) return _current_tick += ticks, true;
-  return false;
-}
-
-bool EventLoop::Scheduler::scheduled(u8 index) const { return _scheduled.test(index); }
-
-void EventLoop::Scheduler::schedule(u8 index, u64 delay) {
-  auto tick = current_tick() + delay;
-  if (_scheduled[index]) [[unlikely]]
-    throw std::runtime_error("Event index is already scheduled");
-  //  Create a new scheduled event in-place at the tail end of the scheduled queue.
-  new (&_queue[_queue_size++]) ScheduledEvent{.tick = tick, .event_index = index};
-  _scheduled.enable_bit(index);
-}
-
-void EventLoop::Scheduler::schedule_over(u8 dependee, u8 dependent, u64 delay) {
-  if (_scheduled[dependee]) [[unlikely]]
-    throw std::runtime_error("Dependee already scheduled");
-
-  // Mark dependent as waiting on dependee
-  _dependencies[dependent] |= index_to_bitmask(dependee);
-  _scheduled[dependent] = false;
-  // Mark dependee as blocking dependent.
-  _dependents[dependee] |= index_to_bitmask(dependent);
-  _scheduled.enable_bit(dependee);
-
-  // Check if the event is currently scheduled for execution.
-  u16 idx = 0;
-  for (; idx < _queue_size; idx++) {
-    if (_queue[idx].event_index == dependent) break;
-  }
-
-  // The dependent is currently scheduled! Rather than pause that event and schedule a new one, just steal its spot
-  if (idx != _queue_size) _queue[idx] = ScheduledEvent{.tick = _current_tick + delay, .event_index = dependee};
-  // Event is not scheduled (already paused?), so we need to allocate a new spot.
-  else new (&_queue[_queue_size++]) ScheduledEvent{.tick = _current_tick + delay, .event_index = dependee};
-}
-
-void EventLoop::Scheduler::pause(u8 dependent, pepp::FixedBitset<MAX_EVENTS> dependees) {
-  if (dependent >= MAX_EVENTS) [[unlikely]]
-    throw std::out_of_range("Dependent index must be less than MAX_EVENTS");
-  else if (dependees.test(dependent)) [[unlikely]] throw std::runtime_error("Event cannot depend on itself");
-
-  _scheduled[dependent] = true;
-
-  // Compute reverse dependencies to speed up hot code in event loop.
-  _dependencies[dependent] |= dependees;
-  for (u64 bits = dependees(); bits;) {
-    u8 dependee = std::countr_zero(bits);
-    bits &= bits - 1;
-    _dependents[dependee].enable_bit(dependent);
-  }
-
-  // Check if the event is currently scheduled for execution. If so, deschedule it.
-  for (u16 idx = 0; idx < _queue_size; idx++) {
-    if (_queue[idx].event_index == dependent) {
-      // Move last item into the hole.
-      _queue[idx] = std::move(_queue[--_queue_size]);
-      if (idx == 0) resort_queue();
-      break;
-    }
-  }
-}
-
-u8 EventLoop::Scheduler::consume_top() {
-  resort_queue();
-  const auto scheduled_idx = _queue[0].event_index;
-  _current_tick = _queue[0].tick, _counters.executed++;
-  // If there are item left in the queue, maintain top-1 sorting requirement
-  if (_queue_size-- > 1) _queue[0] = _queue[_queue_size];
-  _scheduled.clear_bit(scheduled_idx);
-  return scheduled_idx;
-}
-
-void EventLoop::Scheduler::resume(u8 idx) {
-  for (u64 bits = _dependents[idx](); bits;) {
-    u8 paused_idx = std::countr_zero(bits);
-    bits &= bits - 1;
-    _dependencies[paused_idx].reset(idx);
-    if (_dependencies[paused_idx].none()) new (&_queue[_queue_size++]) ScheduledEvent(_current_tick, paused_idx);
-  }
-}
-
-void EventLoop::Scheduler::resort_queue() {
-  if (_queue_size <= 1) return; // No need to sort if we have 0 or 1 events
-  auto end = _queue.begin() + _queue_size;
-  auto min = std::min_element(_queue.begin(), end, [](const auto &a, const auto &b) { return a.tick < b.tick; });
-  std::swap(*min, _queue[0]);
-}
-
 EventLoop::Status EventLoop::run(u64 max_ticks) {
   return run([this, max_ticks] { return scheduler.current_tick() >= max_ticks; });
 }
@@ -147,9 +11,9 @@ EventLoop::Status EventLoop::run(std::function<bool()> pause) {
 
 void EventLoop::dump_state() const {
   fmt::println("{:04x} Current simulation state:", scheduler.current_tick());
-  fmt::println("     Allocated events: {}", scheduler.count());
+  fmt::println("     Allocated events: {}", allocator.current_allocated());
   fmt::println("       paused: {}", paused_events());
-  fmt::println("       scheduled: {}", allocator.count());
+  fmt::println("       scheduled: {}", scheduler.current_scheduled());
   fmt::println("     Event Descriptors");
   /*for (int it = 0; it < MAX_EVENTS; it++) {
     if (_event_slots_used.test(it)) {
